+++ tiko-examples/01_basic_di/src/test/java/io/tiko/examples/basic/PickTest.java
@@ -3,7 +3,7 @@ package io.tiko.examples.basic;
 import io.tiko.Container;
 import io.tiko.Pick;
 import io.tiko.Provider;
-import io.tiko.Tiko;
+import io.tiko.runtime.Tiko;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;