+++ tiko-runtime/src/main/java/io/tiko/runtime/LocalEventBus.java
@@ -1,4 +1,4 @@
-package io.tiko.event.local;
+package io.tiko.runtime;
 
 import io.tiko.EventBus;
 import io.tiko.EventCallback;